requirements.txt
requests
Pillow
cairosvg
pytz